+++ app/Http/Controllers/PembayaranController.php
@@ -5,6 +5,7 @@
     use App\Models\Pembayaran;
     use Illuminate\Http\Request;
     use Illuminate\Support\Facades\Auth;
+    use Illuminate\Support\Facades\Gate;
 
     class PembayaranController extends Controller
     {
@@ -21,6 +22,15 @@
         
         public function index()
         {
+            // Authorization Admin
+            if(Gate::denies('admin')){
+                return response()->json([
+                    'success' => false,
+                    'status' => 403,
+                    'message' => 'you are unauthorized'
+                ], 403);
+            }
+
             $pembayaran = Pembayaran::with('booking')->OrderBy("id", "DESC")->paginate(10)->toArray();
             $response = [
                 "total_count" => $pembayaran["total"],
@@ -45,6 +55,14 @@
         
         public function destroy($id)
         {
+            // Authorization Admin
+            if(Gate::denies('admin')){
+                return response()->json([
+                    'success' => false,
+                    'status' => 403,
+                    'message' => 'you are unauthorized'
+                ], 403);
+            }
             $pembayaran = Pembayaran::find($id);
             $pembayaran->delete();
             $message = ['message' => 'delete sucessfull', 'id' => $id ];
